Moves LegacyPresentation component to components/elements folder
