+++ pages/ignis.vue
@@ -92,7 +92,7 @@ for rb_seed,rb_circ_seed in enumerate(rb_circs):
 <script lang="ts">
 import { Component } from 'vue-property-decorator'
 import QiskitElementPage from '~/components/qiskit/QiskitElementPage.vue'
-import LegacyPresentation from '~/components/headers/LegacyPresentation.vue'
+import LegacyPresentation from '~/components/elements/LegacyPresentation.vue'
 import LegacySection from '~/components/sections/LegacySection.vue'
 import SoftwareStack from '~/components/qiskit/SoftwareStack.vue'
 import SyntaxHighlight from '~/components/qiskit/SyntaxHighlight.vue'
