+++ pages/index.vue
@@ -146,7 +146,7 @@
 <script lang="ts">
 import { Component } from 'vue-property-decorator'
 import QiskitPage from '~/components/qiskit/QiskitPage.vue'
-import LegacyPresentation from '~/components/headers/LegacyPresentation.vue'
+import LegacyPresentation from '~/components/elements/LegacyPresentation.vue'
 import LegacySection from '~/components/sections/LegacySection.vue'
 import Cta from '~/components/ui/Cta.vue'
